Adds an interface for the Board type
